fix(module-ee): widen cost_credits to bigint so the sweep cannot overflow
`ee_usage_records.cost_credits` is the CUMULATIVE credit total for a usage
context, and the durable `unattributed` bucket is one row per org that never
resets. As an `integer` it topped out at 2 147 483 647 credits (~$2.15M).

The overflow was not one org's problem. The sweep writes the column as
`round((cost_usd + delta) * 1000)::int` and derives the org debit from a
RETURNING expression over the same cast, both INSIDE the transaction that
claims the ledger rows: an `integer out of range` there rolled the whole pass
back — no row claimed, no watermark advanced — so billing stopped fleet-wide,
on every tick, until someone found the offending row.

Widen the column to `bigint` (EE migration 0007) and both casts to `::bigint`.
`mode: "number"` keeps the TS type a plain number, exact to 2^53 credits
(~$9e12); the RETURNING delta already goes through `Number()`, which is what
postgres.js needs for an int8 rendered as text.

int4 => int8 is a table rewrite, so this lands while every stored value still
fits in four bytes. No DML, no data repair: the USING conversion is implicit
and lossless.

+++ packages/module-ee/src/billing/usage-recorder.ts
@@ -407,6 +407,13 @@ export async function billLedgerRows(
   // `unattributed` bucket grows without bound, so a float column's absolute
   // error would grow with it.)
   //
+  // WIDTH: `cost_credits` is `bigint` and both casts below are `::bigint`. The
+  // cumulative grows without bound on the durable `unattributed` bucket, and an
+  // `integer out of range` raised here would abort the CALLER's transaction —
+  // one org's overflowing row would stop the sweep for every tenant. See the
+  // column's note in drizzle/schema.ts. The RETURNING delta comes back as a
+  // string (postgres.js renders int8 as text), hence the `Number()` below.
+  //
   // ROUNDING INVARIANT: exactly ONE rounding rule — half away from zero — is
   // shared by `dollarsToCredits` in src/credits.ts (JS `Math.round`, which is
   // half-away-from-zero for the non-negative dollars we bill) and every `round()`
@@ -429,11 +436,11 @@ export async function billLedgerRows(
         target: [orgUsageRecords.contextType, orgUsageRecords.contextId],
         set: {
           costUsd: sql`${orgUsageRecords.costUsd} + ${dollars}`,
-          costCredits: sql`round((${orgUsageRecords.costUsd} + ${dollars}) * ${CREDITS_PER_DOLLAR})::int`,
+          costCredits: sql`round((${orgUsageRecords.costUsd} + ${dollars}) * ${CREDITS_PER_DOLLAR})::bigint`,
         },
       })
       .returning({
-        deltaCredits: sql<number>`${orgUsageRecords.costCredits} - round((${orgUsageRecords.costUsd} - ${dollars}) * ${CREDITS_PER_DOLLAR})::int`,
+        deltaCredits: sql<number>`${orgUsageRecords.costCredits} - round((${orgUsageRecords.costUsd} - ${dollars}) * ${CREDITS_PER_DOLLAR})::bigint`,
       });
     const deltaCredits = Number(row!.deltaCredits);
     if (deltaCredits > 0) perOrg.set(b.orgId, (perOrg.get(b.orgId) ?? 0) + deltaCredits);
